无重大更新

+++ Project_Headers/speed.h
@@ -15,7 +15,7 @@ void SpeedControl(void);
 extern int csl,csr;
 extern int tsl,tsr;
 extern int targetspeed;
-extern byte cyclespeed,straightspeed;;
+extern int cyclespeed,straightspeed;;
 
 extern double Speed_kp_Left,Speed_ki_Left,Speed_kd_Left;
 extern double Speed_kp_Right,Speed_ki_Right,Speed_kd_Right;	//µç»úPID
